Reorganize into multiple crates
