+++ Cargo.toml
@@ -1,37 +1,9 @@
-[package]
-name = "sage"
-version = "0.4.0"
-authors = ["Michael Lazear <michaellazear92@gmail.com"]
-edition = "2021"
-rust-version = "1.62"
-description = "A very fast proteomics database search tool"
-readme = "README.md"
-license = "MIT"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0"
-async-compression = { version = "0.3", features = ["tokio", "gzip", "zlib"] }
-aws-config = "0.49"
-aws-sdk-s3 = "0.19"
-base64 = "0.13"
-bytes = "1.0"
-csv = "1"
-clap = { version="4.0", features = ["cargo", "unicode"] }
-env_logger = "0.8.4"
-log = "0.4.0"
-http = "0.2"
-itoa = "1.0"
-num_cpus = "1.13"
-regex = "1.6"
-rayon = "1.5"
-ryu = "1.0"
-tokio = { version = "1.0", features = ["fs", "io-util", "rt"] }
-tokio-stream = "0.1"
-serde = { version="1.0", features = ["derive"] }
-serde_json = "1.0"
-quick-xml = { version = "0.25", features = ["async-tokio"] }
+[workspace]
+members = [
+    "crates/sage",
+    "crates/sage-cli",
+    "crates/sage-cloudpath",
+]
 
 [profile.release]
 lto = "fat"
