+++ crates/sage-cli/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "sage-cli"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+sage-core = { path = "../sage" }
+sage-cloudpath = { path = "../sage-cloudpath" }
+
+anyhow = "1.0"
+csv = "1"
+clap = { version="4.0", features = ["cargo", "unicode"] }
+env_logger = "0.8.4"
+log = "0.4.0"
+itoa = "1.0"
+num_cpus = "1.13"
+rayon = "1.5"
+ryu = "1.0"
+serde = { version="1.0", features = ["derive"] }
+serde_json = "1.0"
