+++ crates/sage/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "sage-core"
+version = "0.4.0"
+authors = ["Michael Lazear <michaellazear92@gmail.com"]
+edition = "2021"
+rust-version = "1.62"
+description = "A very fast proteomics database search tool"
+readme = "README.md"
+license = "MIT"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+base64 = "0.13"
+csv = "1"
+log = "0.4.0"
+flate2 = "1.0"
+regex = "1.6"
+rayon = "1.5"
+serde = { version="1.0", features = ["derive"] }
+serde_json = "1.0"
+quick-xml = "0.25"
